Minor tidy ups.
